backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai==0.8.0
ffmpeg-python==0.2.0
httpx==0.26.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
